New version of dpdMaker
Improved S1 (4 us and m2) and S2 calculations, additional CRT info for cuts, ped_end update

+++ LIB/gains.h
@@ -1,4 +1,4 @@
-void get_gains(float gains[5], float voltage[5])
+void get_gains(double gains[5], double voltage[5])
 {
 	std::vector<TF1*> _myfit;
